+++ src/helpers/pickObjectKeys.ts
@@ -1,5 +1,12 @@
 import { ObjectKeys, PlainObject } from "../types";
 
+/**
+ *
+ * @description Given an object, returns a new object with only the keys that are in the `keys` array.
+ * @example
+ * const obj = { a: 1, b: 2, c: 3 };
+ * pickObjectKeys(obj, ["a", "c"]); // { a: 1, c: 3 }
+ */
 export const pickObjectKeys = <T extends PlainObject>(
   obj: T,
   keys: ObjectKeys<T>