+++ src/helpers/seriesAsync.ts
@@ -1,7 +1,9 @@
 import { isFunction } from "../validators";
 
-type SeriesResult<T extends readonly unknown[]> = {
-  [K in keyof T]: T[K] extends () => infer U ? Awaited<U> : never;
+type AsyncFunction<T = any> = () => Promise<T>;
+
+type SeriesResult<T extends readonly AsyncFunction[]> = {
+  [K in keyof T]: T[K] extends AsyncFunction<infer U> ? Awaited<U> : never;
 };
 
 /**
@@ -15,16 +17,19 @@ type SeriesResult<T extends readonly unknown[]> = {
  *  async () => 4,
  * ]); => [1, 2, 3, 4]
  */
-export const seriesAsync = async <T extends readonly unknown[]>(
+export const seriesAsync = async <T extends readonly AsyncFunction[]>(
   series: readonly [...T]
 ): Promise<SeriesResult<T>> => {
   const results: unknown[] = [];
   for (const fn of series) {
     if (isFunction(fn)) results.push(await fn());
-    else
+    else {
+      // The error is thrown deliberately because the function will work exactly the same but not being as performant
+      // as it should, because the promises are already executed
       throw new Error(
         `seriesAsync: invalid type received "${typeof fn}", make sure all items are functions, not promises, otherwise they would've been executed already`
       );
+    }
   }
   return results as SeriesResult<T>;
 };