+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # deverything
 
+## 3.0.0
+
+### Major Changes
+
+- getKeys => returns normal keys, like Object.keys
+- groupBy => groupByKey
+- namedSeriesAll => DELETED
+
 ## 2.2.0
 
 ### Minor Changes